# oracle/src/main.rs
use actix_web::{get, post, web, App, HttpResponse, HttpServer};
use serde::{Deserialize, Serialize};
use chrono::Utc;
use reqwest::Client;

async fn call_kyc_oracle(user_id: &str, documents: &[String]) -> Option<KYCResult> {
    Some(KYCResult { status: "verified".to_string(), details: Some("ID verified".to_string()) })
}
async fn fetch_surebits_signature(data: &str) -> Option<String> { Some("surebits_sig_123".to_string()) }
async fn fetch_chainlink_signature(data: &str) -> Option<String> { Some("chainlink_sig_456".to_string()) }
async fn verify_signature(pubkey: &str, signature: &str, data: &str) -> bool { true }
async fn generate_lightning_invoice(amount: &str, dlc: &str) -> Option<String> { Some("lnbc1...".to_string()) }

struct DLCOracle {
    pubkey: String,
    signature: String,
}

async fn attest_ledger_event(data: &str, oracles: Vec<DLCOracle>) -> Option<String> {
    let valid = oracles.iter().filter(|o| verify_signature(&o.pubkey, &o.signature, data)).count() >= 2;
    if valid {
        Some("tx-attested-123".to_string())
    } else {
        None
    }
}

#[derive(Deserialize)]
struct KYCRequest {
    user_id: String,
    documents: Vec<String>,
}

#[derive(Serialize)]
struct KYCResponse {
    status: String,
    details: String,
}

#[derive(Deserialize)]
struct PaymentRequest {
    user_id: String,
    amount: String,
}

#[derive(Serialize)]
struct PaymentResponse {
    invoice: String,
}

struct KYCResult {
    status: String,
    details: Option<String>,
}

#[post("/identity/kyc")]
async fn verify_kyc(req: web::Json<KYCRequest>) -> HttpResponse {
    let kyc_result = call_kyc_oracle(&req.user_id, &req.documents).await.unwrap_or(KYCResult {
        status: "pending".to_string(),
        details: None,
    });
    HttpResponse::Ok().json(KYCResponse {
        status: kyc_result.status,
        details: kyc_result.details.unwrap_or("Pending verification".to_string()),
    })
}

#[post("/billing/pay")]
async fn process_payment(req: web::Json<PaymentRequest>) -> HttpResponse {
    let oracles = vec![
        DLCOracle { pubkey: "surebits_key", signature: fetch_surebits_signature(&req.user_id).await.unwrap_or_default() },
        DLCOracle { pubkey: "chainlink_key", signature: fetch_chainlink_signature(&req.user_id).await.unwrap_or_default() },
    ];
    let dlc = attest_ledger_event(&req.amount, oracles).await.unwrap_or("dlc-123".to_string());
    let invoice = generate_lightning_invoice(&req.amount, &dlc).await.unwrap_or_default();
    HttpResponse::Ok().json(PaymentResponse { invoice })
}

#[get("/health")]
async fn health() -> HttpResponse {
    HttpResponse::Ok().body("Oracle service running")
}

#[actix_web::main]
async fn main() -> std::io::Result<()> {
    HttpServer::new(|| {
        App::new()
            .service(verify_kyc)
            .service(process_payment)
            .service(health)
    })
    .bind("0.0.0.0:3003")?
    .run()
    .await
}